Input validation on client side

+++ server/controllers/article.js
@@ -19,46 +19,59 @@ exports.newArticle = [
   async (req, res) => {
     console.log(req.file)
     console.log(req.body)
-    const s3 = new aws.S3()
+    
     try {
       const { title, content, userId } = req.body
       // const imageUrl = './images/default-article.jpg'
       const revContent = content.split('"').join("'")
-      const url =
-        Math.round(Math.random() * 1000000) +
-        '-' +
-        title
-          .toLowerCase()
-          .split(' ')
-          .join('-')
       const date = Date.now()
 
-      const buffer = await sharp(req.file.path)
-        .resize(950)
-        .toBuffer()
+      // If user uploaded a cover picture, we save the article including the cover image
+      if (req.file) {
+        const buffer = await sharp(req.file.path)
+          .resize(950)
+          .toBuffer()
 
-      const s3res = await s3
-        .upload({
-          Bucket: 'manu2711groupomania/cover',
-          Key: `${Date.now() + '-' + req.file.originalname}`,
-          Body: buffer,
-          ACL: 'public-read'
+        const s3 = new aws.S3()
+        const s3res = await s3
+          .upload({
+            Bucket: 'manu2711groupomania/cover',
+            Key: `${Date.now() + '-' + req.file.originalname}`,
+            Body: buffer,
+            ACL: 'public-read'
+          })
+          .promise()
+        // Connection to Database
+        const conn = await pool.getConnection()
+
+        // Save new article inside database
+        await conn.query(
+          `INSERT INTO articles VALUES (NULL, "${title}", "${revContent}", "${date}", "${userId}", "${s3res.Location}")`
+        )
+        fs.unlink(req.file.path, () => {
+          console.log(s3res.Location)
         })
-        .promise()
-      // Connection to Database
-      const conn = await pool.getConnection()
+        res.status(201).send({
+          message: `Thanks for sharing your new article: ${title} ! `
+        })
+        conn.release()
+      }
 
-      // Save new article inside database
-      await conn.query(
-        `INSERT INTO articles VALUES (NULL, "${title}", "${revContent}", "${date}", "${userId}", "${url}", "${s3res.Location}")`
-      )
-      fs.unlink(req.file.path, () => {
-        console.log(s3res.Location)
-      })
-      res.status(201).send({
-        message: `Thanks for sharing your new article: ${title} ! avec l'url: ${url}`
-      })
-      conn.release()
+      if (!req.file) {
+        const articleCoverDefault = 'https://manu2711groupomania.s3.eu-west-3.amazonaws.com/cover/default_cover.png'
+        // Connection to Database
+        const conn = await pool.getConnection()
+
+        // Save new article inside database
+        await conn.query(
+          `INSERT INTO articles VALUES (NULL, "${title}", "${revContent}", "${date}", "${userId}", "${articleCoverDefault}")`
+        )
+        res.status(201).send({
+          message: `Thanks for sharing your new article: ${title} ! `
+        })
+        conn.release()
+      }
+
     } catch (error) {
       res.status(500).json({ error })
       console.log(error)
@@ -94,10 +107,10 @@ exports.oneArticle = async (req, res) => {
       return res.send({ message: 'There is no article with that id !' })
     }
     const comments = await conn.query(`
-    SELECT comments.content FROM comments
-    INNER JOIN articles
-    ON comments.article_id = articles.id
-    WHERE articles.url="${articleId}"
+    SELECT comments.content, users.name FROM comments
+    INNER JOIN articles ON comments.article_id = articles.id
+    INNER JOIN users ON comments.user_id = users.id
+    WHERE articles.id="${articleId}"
     `)
     res.status(200).send({ article, comments })
     conn.end()
@@ -130,15 +143,45 @@ exports.editArticle = [
     try {
       const articleId = req.params.id
       const { title, content } = req.body
+
+      if (req.file) {
+        const buffer = await sharp(req.file.path)
+          .resize(950)
+          .toBuffer()
+
+        const s3 = new aws.S3()
+        const s3res = await s3
+          .upload({
+            Bucket: 'manu2711groupomania/cover',
+            Key: `${Date.now() + '-' + req.file.originalname}`,
+            Body: buffer,
+            ACL: 'public-read'
+          })
+          .promise()
+        // Connection to Database
+        const conn = await pool.getConnection()
+
+        // Save new article inside database
+        await conn.query(
+          `UPDATE articles SET title="${title}", content="${content}", image_url="${s3res.Location}" WHERE id="${articleId}"`
+        )
+        fs.unlink(req.file.path, () => {
+          console.log(s3res.Location)
+        })
+        res.status(201).send({
+          message: `Thanks for sharing your new article: ${title} ! `
+        })
+        conn.release()
+      }
+
       if (!req.file) {
         const conn = await pool.getConnection()
         await conn.query(
-          `UPDATE articles SET title='${title}', content="${content}", url='abc' WHERE id='${articleId}'`
+          `UPDATE articles SET title='${title}', content="${content}" WHERE id='${articleId}'`
         )
         res.status(200).json({ message: 'Your article has been updated !' })
         conn.release()
       }
-      console.log(req.file)
     } catch (error) {
       res.status(500).json({ error })
       console.log(error)
